2.0.0 Release

+++ BenchmarkCmp/Models/Input/Json/BenchmarkMemory.cs
@@ -1,7 +1,7 @@
 ﻿namespace BenchmarkCmp.Models.Input.Json
 {
     public record BenchmarkMemory(
-        long BytesAllocatedPerOperation,
+        double BytesAllocatedPerOperation,
         double Gen0Collections,
         double Gen1Collections,
         double Gen2Collections);
